// src/app/interfaces/membership.ts
import { GroupOwnership } from "./group-ownership";
import { User } from "./user";

export interface Membership {
  uid: string;
  userUid: string;
  groupUid: string;
  memberStatus: 'ACTIVE' | 'INACTIVE' | 'PENDING' | 'SUSPENDED';
  createdAt?: string;
  updatedAt?: string;
  user: User;
  active: boolean;
  group: GroupOwnership;
}

export interface MembershipCreateDto {
  userUid: string;
  groupUid: string;
  memberStatus: 'ACTIVE' | 'INACTIVE' | 'PENDING' | 'SUSPENDED';
  uid?: string; // Optional for create, will be generated by server
}

export interface MembershipUpdateDto extends Partial<MembershipCreateDto> {}